refactor(model): decouple winner logic using WinnerStrategy
