+++ src/main/kotlin/racingcar/controller/RaceController.kt
@@ -3,11 +3,15 @@ package racingcar.controller
 import racingcar.model.Cars
 import racingcar.view.InputView
 import racingcar.view.OutputView
-import racingcar.strategy.MoveStrategy
-import racingcar.strategy.RandomMoveStrategy
+import racingcar.strategy.move.MoveStrategy
+import racingcar.strategy.move.RandomMoveStrategy
+import racingcar.strategy.winner.WinnerStrategy
+import racingcar.strategy.winner.MaxPositionStrategy
+
 
 class RaceController(
-    private val moveStrategy: MoveStrategy = RandomMoveStrategy()
+    private val moveStrategy: MoveStrategy = RandomMoveStrategy(),
+    private val winnerStrategy: WinnerStrategy = MaxPositionStrategy()
 ) {
     fun play() {
         val names=InputView.enterCarNames()
@@ -21,7 +25,7 @@ class RaceController(
             OutputView.printRoundResult(cars.getAll())
         }
 
-        val winners=cars.findWinners()
+        val winners=cars.findWinners(winnerStrategy)
         OutputView.printWinners(winners)
     }
 }
