+++ src/main/kotlin/racingcar/strategy/move/MoveStrategy.kt
@@ -1,4 +1,4 @@
-package racingcar.strategy
+package racingcar.strategy.move
 
 fun interface MoveStrategy {
     fun shouldMove(): Boolean
