+++ src/main/kotlin/racingcar/strategy/winner/WinnerStrategy.kt
@@ -1,0 +1,7 @@
+package racingcar.strategy.winner
+
+import racingcar.model.Car
+
+fun interface WinnerStrategy {
+    fun selectWinners(cars: List<Car>): List<Car>
+}
